Guard against uninitialized dmah_entry in qede_tx_mapped()
The compiler determined that we were using the dmah_entry variable
potentially uninitialized. See, for example:

cc1: warnings being treated as errors
../../common/io/qede/qede_fp.c: In function 'qede_tx_mapped':
../../common/io/qede/qede_fp.c:1036: error: 'dmah_entry' may be used
uninitialized in this function [-Wuninitialized]

What I noticed was that this could happen if the series of mblk_t's we
got led us to basically never end up processing. This could happen
because of the check around MBLKL at the start of the loop processing
the mblk.