Fix uninitialized variable warnings in qede_osal.c.
The problem here is simple. We literally declare a variable on the stack
and return it. Thankfully no one calls this at this moment, but it seems
kind of pointless.

cc1: warnings being treated as errors
../../common/io/qede/qede_osal.c: In function 'qede_osal_valloc':
../../common/io/qede/qede_osal.c:721: error: 'ptr' is used uninitialized
in this function [-Wuninitialized]
*** Error code 1
make: Fatal error: Command failed for target `debug64/qede_osal.o'